serverless/ses-mail/requirements.txt
attrs==19.3.0
boto==2.49.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.4.5.1
chardet==3.0.4
cleo==0.7.6
clikit==0.4.3
html5lib==1.0.1
idna==2.9
Jinja2==2.11.2
jsonschema==3.2.0
keyring==20.0.1
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==1.0.0
pastel==0.2.0
pexpect==4.8.0
pkginfo==1.5.0.1
poetry==1.0.5
ptyprocess==0.6.0
pydantic==1.7.3
pylev==1.3.0
pyparsing==2.4.7
pyrsistent==0.14.11
requests==2.23.0
requests-toolbelt==0.8.0
shellingham==1.3.2
six==1.14.0
tomlkit==0.5.11
urllib3==1.25.9
webencodings==0.5.1
